+++ src/settings/config.rs
@@ -3,18 +3,26 @@ use dirs;
 use serde::{Deserialize, Serialize};
 use std::fs;
 use std::path::{Path, PathBuf};
+use structopt::StructOpt;
 use toml;
 use tracing::debug;
 
-#[derive(Default, Serialize, Deserialize)]
+#[derive(Default, Serialize, Deserialize, Clone, Debug)]
 pub struct Config {
     pub auto_commit: Option<bool>,
     pub ui: Option<UiTomlTable>,
 }
 
-#[derive(Default, Serialize, Deserialize)]
+#[derive(Default, Serialize, Deserialize, Clone, Debug)]
 pub struct UiTomlTable {
     pub prompt: Option<String>,
+    pub edit_mode: Option<EditMode>,
+}
+
+#[derive(StructOpt, Serialize, Deserialize, Clone, Debug)]
+pub enum EditMode {
+    Emacs,
+    Vi,
 }
 
 impl Config {
@@ -42,6 +50,10 @@ impl Config {
             );
             Ok(Config::default())
         } else {
+            debug!(
+                path = config_file.display().to_string().as_str(),
+                "Loading config"
+            );
             Config::load(&config_file)
         }
     }